refactor(network): split core into submodules
