+++ fold_node/src/network/core.rs
@@ -1,17 +1,13 @@
 use crate::network::config::NetworkConfig;
-use crate::network::error::{NetworkError, NetworkResult};
+use crate::network::error::NetworkResult;
 use crate::network::schema_protocol::SCHEMA_PROTOCOL_NAME;
 use crate::network::schema_service::SchemaService;
-use crate::network::{connections, discovery};
 use libp2p::PeerId;
-use serde_json::Value;
 use std::collections::{HashMap, HashSet};
-use std::time::Duration;
-use log::{info, warn, error};
+use log::info;
 
 /// Core network component for P2P communication between DataFold nodes.
 ///
-/// NetworkCore manages peer connections, discovery, and message routing in the
 /// DataFold network. It provides the foundation for distributed operations across
 /// multiple nodes.
 ///
@@ -46,27 +42,27 @@ use log::{info, warn, error};
 /// ```
 pub struct NetworkCore {
     /// Schema service for handling schema operations
-    schema_service: SchemaService,
+    pub(crate) schema_service: SchemaService,
     /// Local peer ID
-    local_peer_id: PeerId,
+    pub(crate) local_peer_id: PeerId,
     /// Known peers
-    known_peers: HashSet<PeerId>,
+    pub(crate) known_peers: HashSet<PeerId>,
     /// Request timeout in seconds
     #[allow(dead_code)]
-    request_timeout: u64,
+    pub(crate) request_timeout: u64,
     /// Network configuration
-    config: NetworkConfig,
+    pub(crate) config: NetworkConfig,
     /// Mapping from node IDs (UUIDs) to peer IDs
-    node_to_peer_map: HashMap<String, PeerId>,
+    pub(crate) node_to_peer_map: HashMap<String, PeerId>,
     /// Mapping from peer IDs to node IDs (UUIDs)
-    peer_to_node_map: HashMap<PeerId, String>,
+    pub(crate) peer_to_node_map: HashMap<PeerId, String>,
     /// Mapping from node IDs to their listening addresses
-    node_to_address_map: HashMap<String, String>,
+    pub(crate) node_to_address_map: HashMap<String, String>,
     /// Mock for testing - maps peer IDs to schema services
     #[cfg(test)]
-    mock_peers: HashMap<PeerId, SchemaService>,
+    pub(crate) mock_peers: HashMap<PeerId, SchemaService>,
     /// Handle for the background networking task
-    mdns_handle: Option<tokio::task::JoinHandle<()>>,
+    pub(crate) mdns_handle: Option<tokio::task::JoinHandle<()>>,
 }
 
 impl NetworkCore {
@@ -90,31 +86,6 @@ impl NetworkCore {
         })
     }
 
-    /// Register a node ID with a peer ID
-    pub fn register_node_id(&mut self, node_id: &str, peer_id: PeerId) {
-        self.node_to_peer_map.insert(node_id.to_string(), peer_id);
-        self.peer_to_node_map.insert(peer_id, node_id.to_string());
-    }
-
-    /// Register the listening address for a node ID
-    pub fn register_node_address(&mut self, node_id: &str, address: String) {
-        self.node_to_address_map.insert(node_id.to_string(), address);
-    }
-
-    /// Get the listening address for a node ID
-    pub fn get_address_for_node(&self, node_id: &str) -> Option<String> {
-        self.node_to_address_map.get(node_id).cloned()
-    }
-
-    /// Get the peer ID for a node ID
-    pub fn get_peer_id_for_node(&self, node_id: &str) -> Option<PeerId> {
-        self.node_to_peer_map.get(node_id).cloned()
-    }
-
-    /// Get the node ID for a peer ID
-    pub fn get_node_id_for_peer(&self, peer_id: &PeerId) -> Option<String> {
-        self.peer_to_node_map.get(peer_id).cloned()
-    }
 
     /// Get the local peer ID
     pub fn local_peer_id(&self) -> PeerId {
@@ -199,308 +170,4 @@ impl NetworkCore {
         }
     }
 
-    /// Check which schemas are available on a remote peer.
-    ///
-    /// This function sends a request to a remote peer to check which schemas
-    /// from the provided list are available on that peer. It returns a subset
-    /// of the input schema names that are available on the remote peer.
-    ///
-    /// # Arguments
-    ///
-    /// * `peer_id` - The ID of the peer to check
-    /// * `schema_names` - A list of schema names to check for availability
-    ///
-    /// # Returns
-    ///
-    /// A `NetworkResult` containing a vector of available schema names.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `NetworkError` if:
-    /// * The peer is not found in the known peers list
-    /// * There is a connection error when contacting the peer
-    /// * There is a protocol error in the request/response
-    ///
-    /// # Examples
-    ///
-    /// ```rust,no_run
-    /// use fold_node::network::{NetworkCore, NetworkConfig, NetworkResult};
-    /// use libp2p::PeerId;
-    ///
-    /// #[tokio::main]
-    /// async fn main() -> NetworkResult<()> {
-    ///     let config = NetworkConfig::new("/ip4/0.0.0.0/tcp/9000");
-    ///     let mut network = NetworkCore::new(config).await?;
-    ///     
-    ///     let peer_id = PeerId::random(); // In practice, this comes from discovery
-    ///     let schemas_to_check = vec!["user_profile".to_string(), "posts".to_string()];
-    ///     let available_schemas = network.check_schemas(peer_id, schemas_to_check).await?;
-    ///     println!("Available schemas: {:?}", available_schemas);
-    ///     Ok(())
-    /// }
-    /// ```
-    pub async fn check_schemas(
-        &mut self,
-        peer_id: PeerId,
-        schema_names: Vec<String>,
-    ) -> NetworkResult<Vec<String>> {
-        #[cfg(test)]
-        {
-            // For testing, use the mock peer if available
-            if let Some(peer_service) = self.mock_peers.get(&peer_id) {
-                return Ok(peer_service.check_schemas(&schema_names));
-            }
-        }
-
-        // Check if the peer is known
-        if !self.known_peers.contains(&peer_id) {
-            return Err(NetworkError::ConnectionError(format!(
-                "Peer not found: {}",
-                peer_id
-            )));
-        }
-
-        // This is a placeholder for the actual implementation
-        // In a real implementation, this would:
-        // 1. Create a request message
-        // 2. Send the request to the peer
-        // 3. Wait for the response
-        // 4. Parse and return the response
-
-        // For now, just simulate a response with a random subset of schemas
-        let available_schemas = schema_names
-            .iter()
-            .filter(|_| rand::random::<bool>())
-            .cloned()
-            .collect();
-
-        // Simulate network delay
-        tokio::time::sleep(Duration::from_millis(100)).await;
-
-        Ok(available_schemas)
-    }
-
-    /// Add a known peer to the network
-    pub fn add_known_peer(&mut self, peer_id: PeerId) {
-        self.known_peers.insert(peer_id);
-    }
-
-    /// Get the set of known peers
-    pub fn known_peers(&self) -> &HashSet<PeerId> {
-        &self.known_peers
-    }
-
-    /// Forward a request to another node.
-    ///
-    /// This function forwards a JSON request to another node in the network.
-    /// It handles connection establishment, request serialization, and response
-    /// deserialization.
-    ///
-    /// # Arguments
-    ///
-    /// * `peer_id` - The ID of the peer to forward the request to
-    /// * `request` - The JSON request to forward
-    ///
-    /// # Returns
-    ///
-    /// A `NetworkResult` containing the JSON response from the remote node.
-    ///
-    /// # Errors
-    ///
-    /// Returns a `NetworkError` if:
-    /// * The peer is not found in the known peers list
-    /// * There is a connection error when contacting the peer
-    /// * There is a protocol error in the request/response
-    /// * The response cannot be deserialized
-    ///
-    /// # Examples
-    ///
-    /// ```rust
-    /// use fold_node::network::{NetworkCore, NetworkConfig};
-    /// use libp2p::PeerId;
-    /// 
-    /// # tokio_test::block_on(async {
-    /// # let mut network = NetworkCore::new(NetworkConfig::default()).await?;
-    /// # let peer_id = PeerId::random();
-    /// # network.add_known_peer(peer_id);
-    /// let request = serde_json::json!({
-    ///     "operation": "query",
-    ///     "params": {
-    ///         "schema": "user_profile",
-    ///         "fields": ["username", "email"]
-    ///     }
-    /// });
-    ///
-    /// let response = network.forward_request(peer_id, request).await?;
-    /// println!("Response: {:?}", response);
-    /// # Ok::<(), Box<dyn std::error::Error>>(())
-    /// # });
-    /// ```
-    pub async fn forward_request(
-        &mut self,
-        peer_id: PeerId,
-        request: Value,
-    ) -> NetworkResult<Value> {
-        // Check if the peer is known
-        if !self.known_peers.contains(&peer_id) {
-            return Err(NetworkError::ConnectionError(format!(
-                "Peer not found: {}",
-                peer_id
-            )));
-        }
-
-        // Get the operation type from the request
-        let operation = request
-            .get("operation")
-            .and_then(|v| v.as_str())
-            .ok_or_else(|| {
-                NetworkError::ProtocolError("Missing operation in request".to_string())
-            })?;
-
-        // Get the node ID for this peer if available
-        let node_id = self
-            .get_node_id_for_peer(&peer_id)
-            .unwrap_or_else(|| peer_id.to_string());
-
-        info!(
-            "Forwarding {} request to node {} (peer {})",
-            operation, node_id, peer_id
-        );
-
-        // For now, we'll use a direct TCP connection to the target node
-        // In a real implementation, this would use the libp2p request-response protocol
-
-        // Determine the target node's listening address
-        let target_address = match self.get_address_for_node(&node_id) {
-            Some(addr) => addr,
-            None => {
-                return Err(NetworkError::ConnectionError(format!(
-                    "Address for node {} not found",
-                    node_id
-                )));
-            }
-        };
-
-        info!("Connecting to target node at {}", target_address);
-
-        // Connect to the target node
-        let stream = match tokio::net::TcpStream::connect(&target_address).await {
-            Ok(stream) => stream,
-            Err(e) => {
-                return Err(NetworkError::ConnectionError(format!(
-                    "Failed to connect to target node at {}: {}",
-                    target_address, e
-                )));
-            }
-        };
-
-        // Send the request to the target node
-        let result = connections::send_request_to_node(stream, request.clone()).await;
-
-        match result {
-            Ok(response) => {
-                info!("Received response from target node");
-                Ok(response)
-            }
-            Err(e) => {
-                error!("Error forwarding request to target node: {}", e);
-
-                // If we can't connect to the target node, fall back to simulated responses
-                warn!("Falling back to simulated response");
-
-                match operation {
-                    "query" => {
-                        // Get the schema and fields from the request
-                        let schema = request
-                            .get("params")
-                            .and_then(|v| v.get("schema"))
-                            .and_then(|v| v.as_str())
-                            .ok_or_else(|| {
-                                NetworkError::ProtocolError(
-                                    "Missing schema in query request".to_string(),
-                                )
-                            })?;
-
-                        let fields = request
-                            .get("params")
-                            .and_then(|v| v.get("fields"))
-                            .and_then(|v| v.as_array())
-                            .ok_or_else(|| {
-                                NetworkError::ProtocolError(
-                                    "Missing fields in query request".to_string(),
-                                )
-                            })?;
-
-                        // Return a simulated query result
-                        Ok(serde_json::json!({
-                            "results": [
-                                // Generate a result for each field
-                                fields.iter().map(|_| {
-                                    // Generate a random value based on the field type
-                                    match rand::random::<u8>() % 3 {
-                                        0 => serde_json::json!("sample_string_value"),
-                                        1 => serde_json::json!(42),
-                                        _ => serde_json::json!(true),
-                                    }
-                                }).collect::<Vec<_>>()
-                            ],
-                            "schema": schema,
-                            "forwarded": true,
-                            "node_id": node_id,
-                            "peer_id": peer_id.to_string(),
-                            "simulated": true
-                        }))
-                    }
-                    "mutation" => {
-                        // Get the schema from the request
-                        let schema = request
-                            .get("params")
-                            .and_then(|v| v.get("schema"))
-                            .and_then(|v| v.as_str())
-                            .ok_or_else(|| {
-                                NetworkError::ProtocolError(
-                                    "Missing schema in mutation request".to_string(),
-                                )
-                            })?;
-
-                        // Return a simulated mutation result
-                        Ok(serde_json::json!({
-                            "success": true,
-                            "id": format!("simulated_id_{}", rand::random::<u32>()),
-                            "schema": schema,
-                            "forwarded": true,
-                            "node_id": node_id,
-                            "peer_id": peer_id.to_string(),
-                            "simulated": true
-                        }))
-                    }
-                    _ => {
-                        // For other operations, return a generic response
-                        Ok(serde_json::json!({
-                            "success": true,
-                            "operation": operation,
-                            "forwarded": true,
-                            "node_id": node_id,
-                            "peer_id": peer_id.to_string(),
-                            "message": "Request forwarding simulation",
-                            "simulated": true
-                        }))
-                    }
-                }
-            }
-        }
-    }
-
-
-    /// Actively scan for peers using mDNS
-    pub async fn discover_nodes(&mut self) -> NetworkResult<Vec<PeerId>> {
-        discovery::discover_peers(&self.config, &mut self.known_peers).await
-    }
-
-    /// Add a mock peer for testing
-    #[cfg(test)]
-    pub fn add_mock_peer(&mut self, peer_id: PeerId, schema_service: SchemaService) {
-        self.mock_peers.insert(peer_id, schema_service);
-        self.known_peers.insert(peer_id);
-    }
 }
